pyproject.toml

[tool.poetry]
name = "dig"
version = "0.1.0"
description = ""
authors = ["amitsou <amitsou95@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.4.0"
torchvision = "^0.19.0"
numpy = "^2.1.0"
pillow = "^10.4.0"
opencv-python = "^4.10.0"

[tool.poetry.dev-dependencies]
pre-commit = "^3.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
